Chop `manifest` module into multiple submodules
commit-id:27569d02
